+++ src/Akeneo/Tool/Bundle/StorageUtilsBundle/Doctrine/ORM/MappingsOverrideConfigurator.php
@@ -3,8 +3,8 @@
 namespace Akeneo\Tool\Bundle\StorageUtilsBundle\Doctrine\ORM;
 
 use Akeneo\Tool\Bundle\StorageUtilsBundle\Doctrine\MappingsOverrideConfiguratorInterface;
-use Doctrine\Common\Persistence\Mapping\ClassMetadata;
 use Doctrine\ORM\Configuration;
+use Doctrine\ORM\Mapping\ClassMetadata;
 use Doctrine\ORM\Mapping\ClassMetadata as OrmClassMetadata;
 use Doctrine\ORM\Mapping\ClassMetadataInfo;